Move more helper classes to the common->internal->helper module
COSDK-500
