+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/DefaultAnalyticsManager.kt
@@ -11,7 +11,7 @@ package com.adyen.checkout.core.analytics.internal
 import androidx.annotation.VisibleForTesting
 import com.adyen.checkout.core.DispatcherProvider
 import com.adyen.checkout.core.analytics.internal.data.AnalyticsRepository
-import com.adyen.checkout.core.internal.util.runSuspendCatching
+import com.adyen.checkout.core.common.internal.helper.runSuspendCatching
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
