+++ core/src/main/java/com/adyen/checkout/core/common/internal/helper/CheckoutConfigurationMarker.kt
@@ -6,7 +6,7 @@
  * Created by ozgur on 2/5/2025.
  */
 
-package com.adyen.checkout.core.internal.util
+package com.adyen.checkout.core.common.internal.helper
 
 import androidx.annotation.RestrictTo
