+++ core/src/main/java/com/adyen/checkout/core/internal/BaseConfigurationBuilder.kt
@@ -11,8 +11,8 @@ package com.adyen.checkout.core.internal
 import android.content.Context
 import com.adyen.checkout.core.Environment
 import com.adyen.checkout.core.analytics.AnalyticsConfiguration
+import com.adyen.checkout.core.common.internal.helper.LocaleUtil
 import com.adyen.checkout.core.data.model.Amount
-import com.adyen.checkout.core.internal.util.LocaleUtil
 import java.util.Locale
 
 @Suppress(
