+++ core/src/main/java/com/adyen/checkout/core/mbway/internal/ui/MBWayDelegate.kt
@@ -12,6 +12,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.adyen.checkout.core.analytics.internal.AnalyticsManager
+import com.adyen.checkout.core.common.internal.helper.bufferedChannel
 import com.adyen.checkout.core.data.OrderRequest
 import com.adyen.checkout.core.internal.PaymentComponentEvent
 import com.adyen.checkout.core.internal.ui.EventDelegate
@@ -20,7 +21,6 @@ import com.adyen.checkout.core.internal.ui.model.ButtonComponentParams
 import com.adyen.checkout.core.internal.ui.state.DelegateStateManager
 import com.adyen.checkout.core.internal.ui.state.FieldChangeListener
 import com.adyen.checkout.core.internal.ui.state.transformer.FieldTransformerRegistry
-import com.adyen.checkout.core.internal.util.bufferedChannel
 import com.adyen.checkout.core.mbway.internal.ui.model.MBWayDelegateState
 import com.adyen.checkout.core.mbway.internal.ui.model.MBWayViewState
 import com.adyen.checkout.core.mbway.internal.ui.model.toViewState
@@ -65,7 +65,7 @@ internal class MBWayDelegate(
             .stateIn(
                 coroutineScope,
                 SharingStarted.Lazily,
-                toComponentState(stateManager.state.value)
+                toComponentState(stateManager.state.value),
             )
     }
